Add Graphite to Tools (#108)
### What changed?

- Added a Table of Contents to improve navigation
- Added Graphite CLI to the Terminal Tools list
- Added Graphite to the new Web Applications section

### Why make this change?

This update enhances the organization and readability of the TOOLS.md
file. The addition of a Table of Contents allows for quick navigation to
specific sections, while the new categories provide better structure for
the listed tools. Including Graphite CLI and Graphite web application
expands the list of useful tools I use.