Add Security Policy (#143)
# Pull Request

## Description

This pull request includes updates to the documentation to improve
security and project organization. The most important changes include
the addition of a new security policy and an update to the
`.stow-local-ignore` file.

Documentation updates:

*
[`docs/SECURITY.md`](diffhunk://#diff-2605206b7e64be69186fabc8e8fed35676e0015832ff630141aa4aecd73354e0R1-R25):
Added a comprehensive security policy detailing how to report
vulnerabilities, supported versions, and security updates.

Project organization:

*
[`.stow-local-ignore`](diffhunk://#diff-611bebf47e0899a4227c355b253927354186061a26fbecc896decd1b46543fc5R13):
Added `docs` to the list of ignored files.