Add Update Script (#145)
# Pull Request

## Description

This pull request includes updates to the `.zshrc` file and the addition
of a new script for updating and pushing changes. The most important
changes include adding a new alias for the update script and creating
the update script itself.

Updates to `.zshrc`:

* Added an alias `update` to run the `update_and_push.sh` script.

New script addition:

* Created `development-configuration/scripts/update_and_push.sh` to
automate committing and pushing changes to the remote repository.