Update Brewfile (#164)
# Pull Request

## Description

This pull request includes several additions to the `Brewfile` to
enhance the development environment with new tools and utilities.

Additions to `Brewfile`:

* Added `brew "opentofu"` and `brew "ripgrep"` to the list of Homebrew
packages.
[[1]](diffhunk://#diff-1b33d9c13a9a1b0c5d5bc83697ba7b4d36e6cf45f78184c3a28527ffc4418e2dR40)
[[2]](diffhunk://#diff-1b33d9c13a9a1b0c5d5bc83697ba7b4d36e6cf45f78184c3a28527ffc4418e2dR49-R52)
* Added `cask "ghostty"` to the list of Homebrew casks.

Additions to VS Code extensions:

* Added `vscode "42crunch.vscode-openapi"`, `vscode
"alexcvzz.vscode-sqlite"`, and `vscode "bogem.bogems-night-owl"` to the
list of VS Code extensions.
* Added `vscode "hediet.vscode-drawio"` to the list of VS Code
extensions.
* Added `vscode "redhat.vscode-yaml"` and `vscode
"rust-lang.rust-analyzer"` to the list of VS Code extensions.