Upgrade Markdown Link Checker (#175)
# Pull Request

## Description

This pull request includes an update to the
`.github/workflows/check-everything.yml` file to ensure the workflow
uses the latest versions of relevant tools.

Workflow updates:

*
[`.github/workflows/check-everything.yml`](diffhunk://#diff-4dd84dd6da9ad3de39ce8fb6ebb289ee7a656956984dbe3ff166848b6602a171L36-R44):
Changed `runs-on` from `ubuntu-22.04` to `ubuntu-latest` to always use
the latest Ubuntu version.
*
[`.github/workflows/check-everything.yml`](diffhunk://#diff-4dd84dd6da9ad3de39ce8fb6ebb289ee7a656956984dbe3ff166848b6602a171L36-R44):
Updated `UmbrellaDocs/action-linkspector` from version `v1.2.4` to
`v1.2.5` to use the latest version of the action.