Remove TODO.md (#205)
# Pull Request

## Description

This pull request includes a small change to the `TODO.md` file. The
change involves removing the entire content of the file, which included
three tasks related to Git Worktree scripts, using Bat for the help
command, and modifying pull and change branch scripts.

*
[`TODO.md`](diffhunk://#diff-5c6a1301c6b59b30a040d747d065e861d3dd98bde0e5a4356d92d594e9835986L1-L5):
Removed the entire content of the file, which included tasks related to
Git Worktree scripts, using Bat for the help command, and modifying pull
and change branch scripts.