Use Atuin (#222)
# Pull Request

## Description

This pull request adds support for the `atuin` tool across various
configuration files. The most important changes include adding `atuin`
to the Zsh initialization script, the Homebrew `Brewfile`, and the tools
documentation.

### Integration of `atuin`:

*
[`.zshrc`](diffhunk://#diff-4c2d312ff50ee6b26c2cb601fc96a95eceabe4b456831762e5d6caf41b900383R71-R72):
Added initialization command for `atuin` to the Zsh configuration to
enable its history management features.
*
[`Brewfile`](diffhunk://#diff-1b33d9c13a9a1b0c5d5bc83697ba7b4d36e6cf45f78184c3a28527ffc4418e2dR12):
Added `atuin` to the list of Homebrew packages to be installed.
*
[`TOOLS.md`](diffhunk://#diff-806f3d0616e18ff8209e430b5c718c62e4a2f6ad6f2d4b2b2f21ad7c7e9d7f52R45):
Added `atuin` to the list of useful tools in the documentation.