Remove CodeQL Configuration (#231)
# Pull Request

## Description

This pull request makes adjustments to the CodeQL configuration and
workflow to streamline the analysis process by removing an outdated
custom configuration file.

Updates to CodeQL configuration and workflow:

*
[`.github/other-configs/codeql-config.yml`](diffhunk://#diff-b868036a8a443102b2e66591b454b6e9f8d7dc04cca30fc4e4dddb7d680d7979L1-L3):
Removed the `query-filters` section, which excluded the
`actions/unpinned-tag` query. This eliminates the need for a custom
configuration file.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L127):
Removed the `config-file` parameter pointing to the deleted CodeQL
configuration file, simplifying the workflow configuration.