Add fetch all command (#250)
# Pull Request

## Description

This pull request introduces a new script to fetch updates for all Git
repositories in a specified directory and adds a corresponding alias to
the `.zshrc` file for easier access. The changes streamline the workflow
for managing multiple repositories.

### New script addition:

*
[`development-configuration/scripts/fetch_all.sh`](diffhunk://#diff-194cda0d553f923c15722dfb15bda0c69adf8da686108dcaef643b2f570ad707R1-R23):
Added a script that iterates through all subdirectories in a specified
projects folder, checks if they are Git repositories, and fetches
updates for each repository. Includes error handling for non-existent
directories and non-Git subdirectories.

### Alias update:

*
[`.zshrc`](diffhunk://#diff-4c2d312ff50ee6b26c2cb601fc96a95eceabe4b456831762e5d6caf41b900383L107-R107):
Added a new alias `fetch-all` to invoke the `fetch_all.sh` script,
simplifying the process of running the script from the command line.