Add Box Command (#263)
# Pull Request

## Description

This pull request introduces a new alias in the `.zshrc` file and adds a
corresponding script to streamline a common Git workflow. The changes
aim to improve developer productivity by automating tasks.

### New alias and script addition:

*
[`.zshrc`](diffhunk://#diff-4c2d312ff50ee6b26c2cb601fc96a95eceabe4b456831762e5d6caf41b900383R109-R110):
Added a new alias `box` that runs the `box.sh` script, which is located
in the `development-configuration/scripts` directory.
*
[`development-configuration/scripts/box.sh`](diffhunk://#diff-fb277dc979b347fa702b51409e65c5bca0a1216d4c8b78e33b0b35f8774cf422R1-R5):
Introduced a new script that stages all changes in the Git repository
and runs the `update_and_push.sh` script.