Add obsidian (#315)
# Pull Request

## Description

This pull request adds `Obsidian` as a new tool to both the `Brewfile`
and the `TOOLS.md` file, reflecting its inclusion in the user's
preferred tools and applications.

### Updates to tools and applications:

*
[`Brewfile`](diffhunk://#diff-1b33d9c13a9a1b0c5d5bc83697ba7b4d36e6cf45f78184c3a28527ffc4418e2dR91):
Added `cask "obsidian"` to the list of applications managed by Homebrew.
*
[`TOOLS.md`](diffhunk://#diff-806f3d0616e18ff8209e430b5c718c62e4a2f6ad6f2d4b2b2f21ad7c7e9d7f52R53):
Added a link to `Obsidian` under the list of useful tools.