Add fuzzy finder project command (#316)
# Pull Request

## Description

This pull request updates the `.zshrc` file and removes an unused script
to streamline project switching functionality. The most important
changes include replacing the `project` alias with an inline command
using `fzf` and removing the now-obsolete `project_switch.sh` script.

### Updates to `.zshrc`:

* Replaced the `project` alias with an inline command that uses `fzf` to
list directories in `~/Projects/Personal` and navigate to the selected
one. This eliminates the need for the external `project_switch.sh`
script.

### Removal of unused script:

* Deleted the `development-configuration/scripts/project_switch.sh`
script, which is no longer needed due to the updated `project` alias
functionality.