Add pcode command (#317)
# Pull Request

## Description

This pull request updates the `.zshrc` file and documentation to improve
workflow efficiency by introducing a new alias and clarifying existing
commands. The most important changes include the addition of the `pcode`
alias and updates to the `project` alias and its documentation.

### Alias updates:
*
[`.zshrc`](diffhunk://#diff-4c2d312ff50ee6b26c2cb601fc96a95eceabe4b456831762e5d6caf41b900383L108-R109):
Added a new alias `pcode` to navigate to a selected project directory
and open it in Visual Studio Code (`code .`).
*
[`.zshrc`](diffhunk://#diff-4c2d312ff50ee6b26c2cb601fc96a95eceabe4b456831762e5d6caf41b900383L108-R109):
Updated the `project` alias to use double quotes for consistency and
clarified its purpose.

### Documentation updates:
*
[`development-configuration/commands.md`](diffhunk://#diff-b973a06d51f4776a2fc046d251a6474d6858f68e87e01cd2b67fffa71e8050a2L7-R8):
Updated the description of the `project` command to reflect that it
opens the project in the terminal only.
*
[`development-configuration/commands.md`](diffhunk://#diff-b973a06d51f4776a2fc046d251a6474d6858f68e87e01cd2b67fffa71e8050a2L7-R8):
Added documentation for the new `pcode` command, describing its
functionality.